refactor(mech-stats): move buildMechExamineText to mech-stats module
Extract buildMechExamineText from server.ts into src/data/mech-stats.ts so
the examine-text logic lives next to the MECH_STATS table it reads from.
Call site in server.ts updated to import from the new location.

No behavioural change — same output, same truncation guard (84 bytes).
